Separate plant lines template download and upload steps
