+++ app/forms/submissions/population/step05_content_form.rb
@@ -1,0 +1,12 @@
+module Submissions
+  module Population
+    class Step05ContentForm < PlantPopulationForm
+      property :visibility, default: 'private'
+      property :data_owned_by
+      property :data_provenance
+      property :comments
+
+      validates :visibility, inclusion: { in: %w(published private) }
+    end
+  end
+end
