+++ app/views/submissions/steps/population/_step04.html.haml
@@ -1,2 +1,90 @@
-= render partial: "submissions/annotations", locals: { model: PlantPopulation, collapsible: false, form: content }
-= render partial: "submissions/steps/visibility", locals: { form: content }
+.uploaded-plant-lines{ class: "#{'hidden' unless @content.upload}" }
+  %pre.parser-logs.alert.alert-warning.hidden
+  %pre.parser-errors.alert.alert-danger.hidden
+
+  .form-group
+    = label_tag :upload, 'Uploaded file'
+    %div
+      %span.file-name= @content.upload.try(:file_file_name)
+      = delete_submission_upload_button(@content.upload,
+        class: 'btn btn-default delete-plant-lines-upload')
+
+  .info-block.parser-note
+    %p
+      Please review the above parsing logs. If there were any critical errors during the process,
+      the red section will try to give some hints regarding their nature. You may then try to fix
+      the file accordingly, delete the current upload and re-upload the fixed file.
+
+    %p
+      Otherwise, if you think the parsing went properly, use the <b>NEXT</b> button to confirm
+      these changes and proceed with further steps of submission.
+
+  %hr
+
+.info-block
+  Alternatively, you may use the input below to add plant lines to this population.
+  You are able to select one or more lines from the database or, if suitable plant line
+  does not exist in BIP, you are welcome to add it using the <b>Add new plant line</b>
+  form below.
+
+.form-group
+  = content.label :plant_line_list
+  - option_tags = options_for_submission_select(@content.existing_plant_lines, :plant_line_name)
+  - option_tags.concat options_for_submission_select(@content.new_plant_lines, :plant_line_name, id: false)
+  = content.select :plant_line_list, option_tags.html_safe, {}, multiple: true, class: 'plant-line-list form-control', 'data-placeholder': "Select plant line"
+
+%button.btn.btn-default.new-plant-line-for-list{ type: :button } Add new Plant line
+
+.panel.panel-info.new-plant-line-for-list.hidden
+  .panel-heading
+    %h4 Add new <b>Plant line</b>
+
+  .panel-body{"data-dirty-context": "new-plant-line"}
+    .form-group
+      = label_tag :plant_line_name, nil, class: 'required'
+      = text_field_tag :plant_line_name, nil,
+        class: 'form-control required',
+        placeholder: 'Enter new plant line name',
+        autocomplete: 'off'
+
+    .form-group
+      = label_tag :taxonomy_term, nil, class: 'required'
+      = select_tag :taxonomy_term, options_for_select(TaxonomyTerm.names), include_blank: true, class: 'taxonomy-term form-control required', 'data-placeholder': 'Select term'
+
+    .form-group
+      = label_tag :common_name, 'Common name'
+      = text_field_tag :common_name, nil, class: 'form-control', placeholder: "Enter common name"
+
+    .form-group
+      = combo_field_tag :previous_line_name, nil, nil,
+        class: 'previous-line-name',
+        select_placeholder: "Select existing plant line",
+        input_placeholder: "Enter new plant line name"
+
+    .form-group
+      - option_tags = options_for_select(PlantLine.genetic_statuses)
+      = combo_field_tag :genetic_status, nil, option_tags, class: 'genetic-status'
+
+    .form-group
+      = label_tag :plant_variety_name
+      = select_tag :plant_variety_name, [], class: 'plant-variety-name form-control', 'data-placeholder': "Select plant variety"
+
+    .form-group
+      = label_tag :sequence_identifier
+      = text_field_tag :sequence_identifier, nil, class: 'form-control', placeholder: "Enter sequence identifier"
+
+      %small.help-block
+        Provide the identifier of the plant line sequence, in an external database,
+        like #{link_to "SRA", "http://www.ncbi.nlm.nih.gov/sra/"}. Identifiers
+        starting with SR (like SRX1491141) will be automatically recognized as SRA submissions.
+
+    = render partial: "submissions/annotations", locals: { model: PlantLine }
+
+    .action-buttons
+      %button.btn.btn-default.cancel-new-plant-line-for-list{type: 'button'} Cancel
+      %button.btn.btn-primary.add-new-plant-line-for-list{type: 'button'} Confirm
+
+- @content.new_plant_lines.each do |plant_line|
+  %div{ id: "new-plant-line-#{plant_line.plant_line_name.split.join('-').downcase}", class: "new-plant-line-attrs" }
+    - @content.class.new_plant_line_properties.each do |attr|
+      = hidden_field_tag "submission[content][new_plant_lines][][#{attr}]", plant_line.send(attr)
