+++ app/views/submissions/steps/population/_step05.html.haml
@@ -1,0 +1,2 @@
+= render partial: "submissions/annotations", locals: { model: PlantPopulation, collapsible: false, form: content }
+= render partial: "submissions/steps/visibility", locals: { form: content }
